Some changes
Continues efforts on making the hitboxes work. Changed UniversalHitbox
to FighterHitbox in anticipation of eventual ProjectileHitbox class.
Changed Attack from Struct to Class. :octopus:
