+++ Assets/Scripts/Attack.cs
@@ -1,18 +1,46 @@
 ﻿using UnityEngine;
 using System.Collections;
 
-public struct Attack {
+public class Attack {
     /// <summary>
     /// The properties of the attack.
     /// </summary>
-    public AttackData data;
+    public readonly AttackData kData;
     /// <summary>
     /// The team this attack came from.
     /// </summary>
-    public int team;
+    public readonly int kTeam;
     /// <summary>
     /// The id of this attack; not expected to be meaningful beyond
     /// differentiating it with other attacks.
     /// </summary>
-    public int id;
+    public readonly int kId;
+    /// <summary>
+    /// The damage multiplier applied to the attack. Affected when blocked,
+    /// or when using higher-powered characters.
+    /// </summary>
+    public float kDamageMultiplier;
+
+    public Attack(AttackData data, int team, int id, float damageMult = 1.00f) {
+        kData = data;
+        kTeam = team;
+        kId = id;
+        kDamageMultiplier = damageMult;
+    }
+
+    public static bool operator ==(Attack left, Attack right) {
+        return left.kData == right.kData && left.kTeam == right.kTeam && left.kId == right.kId;
+    }
+    public static bool operator !=(Attack left, Attack right) {
+        return left.kData != right.kData || left.kTeam != right.kTeam || left.kId != right.kId;
+    }
+    public override bool Equals (object obj) {
+        if (obj.GetType () == typeof (Attack)) {
+            var atk = (Attack) obj;
+            return kData == atk.kData && kTeam == atk.kTeam && kId == atk.kId;
+        }
+        else {
+            return false;
+        }
+    }
 }
